requirements.txt
langchain
langchain-community
langchain-huggingface
langchain-groq
chromadb
sentence-transformers
streamlit
python-dotenv

# --- for Assignment 2 ---
rouge-score
nltk
scikit-learn
numpy
tqdm